% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondtest.R
\docType{package}
\name{secondtest}
\alias{secondtest}
\title{secondtest: A package with some basic functions.}
\description{
The package includes a sample function and
a sample dataset with their respective documentation.
}
\author{
Shaila Jamal \email{jamals16@mcmaster.ca}
}
